[Xiaomi Vacuum] Fix cleaning time + control level
* Fix wrong total cleaning time
* Changed format total cleaning time to minutes
* Allowed control of vacuum %
* Refactoring
