+++ addons/binding/org.openhab.binding.xiaomivacuum/src/main/java/org/openhab/binding/xiaomivacuum/internal/robot/FanModeType.java
@@ -6,37 +6,25 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.xiaomivacuum.internal;
+package org.openhab.binding.xiaomivacuum.internal.robot;
 
 /**
  * List of Errors
  *
  * @author Marcel Verpaalen - Initial contribution
  */
-public enum StatusType {
+public enum FanModeType {
 
-    UNKNOWN(0, "Unknown"),
-    INITIATING(1, "Initiating"),
-    SLEEPING(2, "Sleeping"),
-    IDLE(3, "Idle"),
-    UNKNOWN4(4, "Unknown Status 4"),
-    CLEANING(5, "Cleaning"),
-    RETURNING(6, "Returning Dock"),
-    UNKNOWN7(7, "Unknown Status 7"),
-    CHARGING(8, "Charging"),
-    CHARGING_ERROR(9, "Charging Error"),
-    PAUSED(10, "Paused"),
-    SPOTCLEAN(11, "Spot cleaning"),
-    ERROR(12, "In Error"),
-    SHUTTING_DOWN(13, "Shutting Down"),
-    UPDATING(14, "Updating"),
-    DOCKING(15, "Docking"),
-    FULL(100, "Full");
+    SILENT(38, "Silent"),
+    STANDARD(60, "Standard"),
+    POWER(77, "Power"),
+    FULL(90, "Full"),
+    CUSTOM(-1, "Custom");
 
     private final int id;
     private final String description;
 
-    StatusType(int id, String description) {
+    FanModeType(int id, String description) {
         this.id = id;
         this.description = description;
     }
@@ -45,13 +33,13 @@ public enum StatusType {
         return id;
     }
 
-    public static StatusType getType(int value) {
-        for (StatusType st : StatusType.values()) {
+    public static FanModeType getType(int value) {
+        for (FanModeType st : FanModeType.values()) {
             if (st.getId() == value) {
                 return st;
             }
         }
-        return UNKNOWN;
+        return CUSTOM;
     }
 
     public String getDescription() {
