+++ addons/binding/org.openhab.binding.xiaomivacuum/src/main/java/org/openhab/binding/xiaomivacuum/internal/robot/VacuumErrorType.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.xiaomivacuum.internal;
+package org.openhab.binding.xiaomivacuum.internal.robot;
 
 /**
  * List of Errors
